introduced Paper check so BQ 3.0.0 only works on paper now.
